requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
tensorflow>=2.12.0
tensorflow-hub>=0.13.0
pillow>=9.5.0
numpy>=1.23.5
librosa>=0.10.1
soundfile>=0.12.1
scikit-learn>=1.3.0
pydub>=0.25.1 